PR adjustments
