+++ src/Data/Managers/GetMetrics/interface.ts
@@ -1,3 +1,5 @@
+import { BackendMetricData } from "@/Slices/Dashboard/Core/Domain";
+
 export interface GetMetrics {
   kind: "GetMetrics";
   startDate: string;
@@ -5,9 +7,9 @@ export interface GetMetrics {
 }
 
 export interface GetMetricsManifest {
-  error: undefined;
-  apiResponse: undefined;
-  data: boolean;
+  error: string;
+  apiData: { data: BackendMetricData };
+  data: BackendMetricData;
   usedData: boolean;
   query: GetMetrics;
 }
