+++ src/Slices/Dashboard/UI/GraphCard.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import {
   Card,
   CardTitle,
@@ -10,93 +10,16 @@ import {
 } from "@patternfly/react-core";
 import styled from "styled-components";
 import { words } from "@/UI";
-import {
-  GraphCardProps,
-  LegendData,
-  Metric,
-  MetricName,
-  StackedMetric,
-} from "../Core/Domain";
+import { GraphCardProps, Metric, MetricName } from "../Core/Domain";
 import { LineChart } from "./Charts/LineChart";
-import { colorTheme } from "./themes";
+import { formatLegendData, formatMetricsToStacked } from "./helper";
 
 export const GraphCard: React.FC<GraphCardProps> = ({
   isStacked,
   timestamps,
   metrics,
 }) => {
-  const [chartState, setChartState] = useState<Metric[]>([]);
-  const [legendDataState, setLegendDataState] = useState<LegendData[]>([]);
-  const [max, setMax] = useState<number>(0);
-
-  useEffect(() => {
-    if (isStacked) {
-      const tempCharState: Metric[] = [];
-      let max = 0;
-      const { data } = metrics as StackedMetric;
-      const base = data.find((object) => object !== null);
-      if (base !== undefined && base !== null) {
-        const keys = Object.keys(base);
-        keys.map((key) => {
-          tempCharState.push({
-            name: key,
-            data: [],
-          });
-        });
-        data.map((object) => {
-          let tempMax = 0;
-          keys.forEach((key, index) => {
-            tempMax += object === null ? 0 : object[key];
-            tempCharState[index].data.push(
-              object === null ? null : object[key]
-            );
-          });
-          if (max < tempMax) {
-            max = tempMax;
-          }
-        });
-      }
-
-      setChartState(tempCharState);
-      setMax(max);
-      setLegendDataState(
-        tempCharState.map(({ name }) => {
-          return {
-            childName: name,
-            name: name.charAt(0).toUpperCase() + name.slice(1),
-            symbol: {
-              fill:
-                colorTheme[name] === undefined
-                  ? colorTheme.default
-                  : colorTheme[name],
-            },
-          };
-        })
-      );
-    } else {
-      setChartState([metrics as Metric]);
-      setMax(
-        (metrics as Metric).data
-          .flat()
-          .map((value) => (value !== null ? value : 0))
-          .sort((a, b) => a - b)[(metrics as Metric).data.flat().length - 1]
-      );
-      setLegendDataState([
-        {
-          childName: metrics.name,
-          name:
-            metrics.name.split(".")[1].charAt(0).toUpperCase() +
-            metrics.name.split(".")[1].slice(1),
-          symbol: {
-            fill:
-              colorTheme[metrics.name] === undefined
-                ? colorTheme.default
-                : colorTheme[metrics.name],
-          },
-        },
-      ]);
-    }
-  }, [metrics, isStacked]);
+  const [formatedMetrics, max] = formatMetricsToStacked(metrics, isStacked);
   return (
     <Card id={`trend-card-${metrics.name}`} component="div" isRounded>
       <CardHeader>
@@ -119,7 +42,7 @@ export const GraphCard: React.FC<GraphCardProps> = ({
         </Flex>
       </CardHeader>
       <CardBody>
-        {chartState.length > 0 && (
+        {(formatedMetrics as Metric[]).length > 0 && (
           <LineChart
             label={words(`dashboard.${metrics.name as MetricName}.label.x`)}
             title={words(`dashboard.${metrics.name as MetricName}.title`)}
@@ -127,10 +50,10 @@ export const GraphCard: React.FC<GraphCardProps> = ({
               `dashboard.${metrics.name as MetricName}.description`
             )}
             isStacked={isStacked}
-            legendData={legendDataState}
+            legendData={formatLegendData(metrics, isStacked)}
             timestamps={timestamps}
-            metrics={chartState}
-            max={max}
+            metrics={formatedMetrics as Metric[]}
+            max={max as number}
           />
         )}
       </CardBody>
